Rename discord logic classes to "services"
